backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.11
langchain-core==0.1.30
langchain-openai==0.1.0
langgraph==0.0.29
openai==1.3.0
PyGithub==2.1.1
httpx==0.25.2
python-dotenv==1.0.0
